requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
celery==5.3.4
redis==5.0.1
google-cloud-storage==2.10.0
pillow==10.1.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
jinja2==3.1.2
python-dateutil==2.8.2
cryptography==41.0.8